+++ app/src/main/java/net/wildfyre/client/viewmodels/ArchiveFragmentViewModel.kt
@@ -7,7 +7,7 @@ import net.wildfyre.client.data.Post
 import net.wildfyre.client.data.repositories.ArchiveRepository
 
 class ArchiveFragmentViewModel(application: Application) : PostsFragmentViewModel(application) {
-    override var itemCount: LiveData<Long> = Transformations.map(ArchiveRepository.superPost) { it.count ?: 0 }
+    override var itemCount: LiveData<Int> = Transformations.map(ArchiveRepository.superPost) { it.count ?: 0 }
     override val items: LiveData<List<Post>> = ArchiveRepository.posts
 
     override fun fetchNextItems() = ArchiveRepository.fetchNextPosts(this)
